+++ spec/aws_instances/aws_spec_api.rb
@@ -1,5 +1,8 @@
 require 'spec_helper'
 
+SERVER_AMI = 'ami-e0b38888'
+WORKER_AMI = 'ami-a8bc87c0'
+
 describe OpenStudio::Aws::Aws do
   context 'create a new instance' do
     before(:all) do
@@ -7,6 +10,8 @@ describe OpenStudio::Aws::Aws do
       @aws = OpenStudio::Aws::Aws.new
 
       @group_id = nil
+
+      FileUtils.rm_f 'ec2_server_key.pem' if File.exist? 'ec2_server_key.pem'
     end
 
     it 'should create a new instance' do
@@ -15,16 +20,17 @@ describe OpenStudio::Aws::Aws do
     end
 
     it 'should create a server' do
-      options = { instance_type: 'm3.medium', image_id: 'ami-3c0fbf54' }
+      options = { instance_type: 'm3.medium', image_id: SERVER_AMI }
 
-      test_pem_file = 'ec2_server_key.pem'
-      FileUtils.rm_f test_pem_file if File.exist? test_pem_file
+      FileUtils.rm_f 'ec2_server_key.pem' if File.exist? 'ec2_server_key.pem'
       FileUtils.rm_f 'server_data.json' if File.exist? 'server_data.json'
 
       @aws.create_server(options)
 
       expect(File.exist?('ec2_server_key.pem')).to be true
       expect(File.exist?('server_data.json')).to be true
+      expect(File.exist?('ec2_worker_key.pem')).to be true
+      expect(File.exist?('ec2_worker_key.pub')).to be true
       expect(@aws.os_aws.server).not_to be_nil
       expect(@aws.os_aws.server.data.availability_zone).to match /us-east-../
 
@@ -37,7 +43,7 @@ describe OpenStudio::Aws::Aws do
     end
 
     it 'should create a 1 worker' do
-      options = { instance_type: 'm3.medium', image_id: 'ami-040ebe6c' }
+      options = { instance_type: 'm3.medium', image_id: WORKER_AMI }
 
       @aws.create_workers(1, options)
 
@@ -47,7 +53,7 @@ describe OpenStudio::Aws::Aws do
     end
 
     it 'should be able to connect a worker to an existing server' do
-      options = { instance_type: 'm3.medium', image_id: 'ami-3c0fbf54' }
+      options = { instance_type: 'm3.medium', image_id: WORKER_AMI }
 
       # will require a new @aws class--but attached to same group_uuid
       @config = OpenStudio::Aws::Config.new
@@ -61,10 +67,6 @@ describe OpenStudio::Aws::Aws do
     it 'should kill running instances' do
       # how to test this?
     end
-
-    after :all do
-      FileUtils.rm_f 'ec2_server_key.pem' if File.exist? 'ec2_server_key.pem'
-    end
   end
 
   context 'upload data to a server' do
@@ -72,7 +74,7 @@ describe OpenStudio::Aws::Aws do
       config = OpenStudio::Aws::Config.new
       aws = OpenStudio::Aws::Aws.new
 
-      options = { instance_type: 'm3.medium', image_id: 'ami-3c0fbf54' }
+      options = { instance_type: 'm3.medium', image_id: SERVER_AMI }
 
       aws.create_server(options)
     end
@@ -135,7 +137,7 @@ describe OpenStudio::Aws::Aws do
       begin
         options = {
           instance_type: 'm3.medium',
-          image_id: 'ami-3c0fbf54',
+          image_id: SERVER_AMI,
           tags: [
             'ci_tests=true',
             'nothing=else',
@@ -182,7 +184,7 @@ describe OpenStudio::Aws::Aws do
         begin
           options = {
             instance_type: 'm3.medium',
-            image_id: 'ami-3c0fbf54', # server AMI
+            image_id: SERVER_AMI,
             tags: [
               'ci_tests=true',
               'ServerOnly=true'
@@ -210,19 +212,4 @@ describe OpenStudio::Aws::Aws do
       end
     end
   end
-
-  # context 'create ebs storage' do
-  #   before(:all) do
-  #     @aws = OpenStudio::Aws::Aws.new
-  #   end
-  #
-  #   it 'should create an EBS volume' do
-  #     options = {instance_type: 'm1.small', image_id: 'ami-29faca40', ebs_volume_size: 128}
-  #     @aws.create_volume()
-  #     @aws.create_server(options)
-  #     expect(@aws.os_aws.server).not_to be_nil
-  #   end
-  #
-  # end
-
 end