+++ spec/openstudio-aws/aws_wrapper_spec.rb
@@ -96,4 +96,16 @@ describe OpenStudioAwsWrapper do
     end
   end
 
+  context 'security groups' do
+    before :all do
+      @aws = OpenStudio::Aws::Aws.new
+    end
+
+    it 'should create security group' do
+      sg = @aws.os_aws.create_or_retrieve_default_security_group
+      pp sg
+      expect(sg).to_not be nil
+    end
+  end
+
 end